backend/pyproject.toml
[project]
name = "anticlockwisesearch-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "nishizhen",email = "nishizhen@gmail.com"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "watchdog (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
